Get Consensus Peak from Replicates with MSPC.

Get Consensus Peak from Replicates with MSPC.

GetConsensusPeak(
  peak.file,
  peak.folder = NULL,
  mspc.path = NULL,
  rep.type = c("bio", "tec"),
  stringency.threshold = 1e-08,
  weak.threshold = 1e-04,
  gamma = 1e-08,
  alpha = 0.05,
  min.overlap.num = 1,
  multiple.intersections = c("Lowest", "Highest"),
  parallelism.degree = 1
)

Arguments

peak.file
Peak files (two or more file: get consensus peak; one file: read directly) obtained from peak caller, eg: MACS2 (without header).

peak.folder
The folder contains peak files. Default: NULL.

mspc.path
MSPC path. Default: NULL (conduct automatic detection).

rep.type
Replicate type, chosen from bio (biological) and tec (technical).

stringency.threshold
Set a threshold on p-values, where peaks with p-value lower than this threshold, are considered stringent. Default: 1e-8.

weak.threshold
Set a threshold on p-values, such that peaks with p-value between this and stringency threshold, are considered weak peaks. Default: 1e-4.

gamma
Set the combined stringency threshold. Peaks with combined p-value below this threshold are confirmed. Default: 1e-8.

alpha
Set the threshold for Benjamini-Hochberg multiple testing correction. Default: 0.05.

min.overlap.num
Set the minimum number of overlapping peaks required before MSPC combines their p-value. Default: 1.

multiple.intersections
When multiple peaks from a sample overlap with a given peak, this argument defines which of the peaks to be considered: the one with lowest p-value, or the one with highest p-value? Chosen from Lowest and Highest. Default: Lowest.

parallelism.degree
Set the number of parallel threads MSPC can utilize simultaneously when processing data. Default: 1.

Value

A dataframe contains all consensus peak.

Examples

# library(ggcoverage)
# peak.file <- system.file("extdata", "ChIP-seq", "consensus.peak", package = "ggcoverage")
# peak.df <- GetConsensusPeak(peak.file = peak.file)