Add Ideogram Annotation to Coverage Plot.

Add Ideogram Annotation to Coverage Plot.

geom_ideogram(
  genome = "hg19",
  mark.color = "red",
  mark.alpha = 0.7,
  mark.line.size = 1,
  add.shadow = TRUE,
  shadow.color = "grey",
  shadow.alpha = 0.7,
  shadow.line.size = 1,
  highlight.centromere = FALSE,
  highlight.color = "green",
  highlight.alpha = 0.7,
  highlight.line.size = 1,
  highlight.shadow.color = "black",
  highlight.shadow.alpha = 0.7,
  highlight.shadow.line.size = 1,
  plot.space = 0.1,
  plot.height = 0.1
)

Arguments

genome
Single specie names, which must be one of the result from ucscGenomes()$db. If missing, will invoke a menu for users to choose from. Default: hg19.

mark.color
The color to mark plot region on ideogram. Default: "red".

mark.alpha
The alpha to mark plot region on ideogram. Default: 0.7.

mark.line.size
The line size to mark plot region on ideogram. Default: 1.

add.shadow
Logical value, whether to add shadow polygon. Default: TRUE.

shadow.color
The color to fill shadow polygon. Default: "grey".

shadow.alpha
The alpha of shadow polygon. Default: 0.7.

shadow.line.size
The line size of shadow polygon. Default: 1.

highlight.centromere
Logical value, whether to highlight centromere region. Default: FALSE.

highlight.color
The color to mark centromere region on ideogram. Default: "green".

highlight.alpha
The alpha to mark centromere region on ideogram. Default: 0.7.

highlight.line.size
The line size to mark centromere region on ideogram. Default: 1.

highlight.shadow.color
The color to fill centromere shadow polygon. Default: "black".

highlight.shadow.alpha
The alpha of centromere shadow polygon. Default: 0.7.

highlight.shadow.line.size
The line size of centromere shadow polygon. Default: 1.

plot.space
Top and bottom margin. Default: 0.1.

plot.height
The relative height of ideogram annotation to coverage plot. Default: 0.2.

Value

Plot.