Create Mass Spectrometry Protein Coverage Plot.

Create Mass Spectrometry Protein Coverage Plot.

ggprotein(
  coverage.file,
  fasta.file,
  protein.id,
  XCorr.threshold = 2,
  confidence = "High",
  contaminant = NULL,
  remove.na = TRUE,
  color = "grey",
  mark.bare = TRUE,
  mark.color = "red",
  mark.alpha = 0.5,
  show.table = TRUE,
  table.position = c("right_top", "left_top", "left_bottom", "right_bottom"),
  table.size = 4,
  table.color = "black",
  range.size = 3,
  range.position = c("in", "out"),
  plot.space = 0.2
)

Arguments

coverage.file
Exported protein coverage file, should be in excel.

fasta.file
Input reference protein fasta file.

protein.id
The protein ID of exported coverage file. This should be unique and in fasta.file.

XCorr.threshold
The cross-correlation threshold. Default: 2.

confidence
The confidence level. Default: High.

contaminant
Whether to remove contaminant peptides. Default: NULL (not remove).

remove.na
Logical value, whether to remove NA value in Abundance column. Default: TRUE.

color
The fill color of coverage plot. Default: grey.

mark.bare
Logical value, whether to mark region where Abundance is zero or NA. Default: TRUE.

mark.color
The color used for the marked region. Default: red.

mark.alpha
The transparency used for the marked region. Default: 0.5.

show.table
Logical value, whether to show coverage summary table. Default: TRUE.

table.position
The position of the coverage summary table, choose from right_top, left_top, left_bottom, right_bottom. Default: right_top.

table.size
The font size of coverage summary table. Default: 4.

table.color
The font color of coverage summary table. Default: black.

range.size
The label size of range text, used when range.position is in. Default: 3.

range.position
The position of y axis range, chosen from in (move y axis in the plot) and out (normal y axis). Default: in.

Value

A ggplot2 object.

Examples

# library(ggcoverage)
# coverage.file <- system.file("extdata", "Proteomics", "MS_BSA_coverage.xlsx", package = "ggcoverage")
# fasta.file <- system.file("extdata", "Proteomics", "MS_BSA_coverage.fasta", package = "ggcoverage")
# protein.id = "sp|P02769|ALBU_BOVIN"
# ggprotein(coverage.file = coverage.file, fasta.file = fasta.file, protein.id = protein.id)